Fix: Agrupamento de funções

+++ src/components/enemies/EnemyTypeTwo.java
@@ -19,7 +19,7 @@ public class EnemyTypeTwo extends Enemy {
 
     public EnemyTypeTwo(int state, double coordinateX, double coordinateY, double speedX, double speedY,
                         double radius, Instant explosionStart, Instant explosionEnd, Instant nextShoot,
-                        double angle, double rotationSpeed, double speed, int spawnX, int countProjectiles, int projectileRadius) {
+                        double angle, double rotationSpeed, double speed, double spawnX, int countProjectiles, int projectileRadius) {
         super(state, coordinateX, coordinateY, speedX, speedY, radius, explosionStart, explosionEnd, nextShoot, angle, rotationSpeed, speed, countProjectiles, projectileRadius);
         this.spawnX = spawnX;
         this.count = 0;
